requirements.txt
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pyodbc==5.3.0
Werkzeug==3.1.5
python-dotenv==1.0.1
gunicorn==23.0.0